# m2u-server/src/generated/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mobilab2u API collections",
    "version": "0.0.1",
    "description": "All API collections"
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": {
    "Authorization": []
  },
  "host": "0.0.0.0:9000",
  "paths": {
...
    "/api/v1/packages": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Create Plan",
        "security": [
          {
            "Authorization": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "serviceType": {
                    "type": "string"
                  },
                  "members": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "hspShare": {
                    "type": "number"
                  },
                  "gpShare": {
                    "type": "number"
                  },
                  "customerShare": {
                    "type": "number"
                  },
                  "mobilabShare": {
                    "type": "number"
                  },
                  "labShare": {
                    "type": "number"
                  },
                  "offerPrice": {
                    "type": "number"
                  }
                }
              },
              "required": [
                "name",
                "description",
                "type",
                "serviceType",
                "price",
                "hspShare",
                "gpShare",
                "customerShare",
                "mobilabShare",
                "labShare",
                "offerPrice"
              ]
            }
          }
        },
        "responses": {
          "200": {
            "message": "Plan created"
          }
        }
      },
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Get all plans",
        "security": [
          {
            "Authorization": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": null
          }
        },
        "responses": {
          "200": {
            "message": "All plans"
          }
        }
      }
    },
    "/api/v1/packages/{id}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "description": "Update Plan",
        "security": [
          {
            "Authorization": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The ID of the service provider"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "serviceType": {
                    "type": "string"
                  },
                  "members": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "offerPrice": {
                    "type": "number"
                  },
                  "hspShare": {
                    "type": "number"
                  },
                  "gpShare": {
                    "type": "number"
                  },
                  "customerShare": {
                    "type": "number"
                  },
                  "mobilabShare": {
                    "type": "number"
                  },
                  "labShare": {
                    "type": "number"
                  }
                }
              },
              "required": [
                "name",
                "description",
                "type",
                "serviceType",
                "price",
                "hspShare",
                "gpShare",
                "customerShare",
                "mobilabShare",
                "labShare",
                "offerPrice"
              ]
            }
          }
        },
        "responses": {
          "200": {
            "message": "Plan created"
          }
        }
      },
      "get": {
        "tags": [
          "Admin",
          "Customer"
        ],
        "description": "Get dingle plan",
        "security": [
          {
            "Authorization": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The ID of the service provider"
          }
        ],
        "responses": {
          "200": {
            "message": "All plans"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "description": "Delete Package",
        "security": [
          {
            "Authorization": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The ID of the Package"
          }
        ],
        "responses": {
          "200": {
            "message": "Delete Packages"
          }
        }
      }
    },
    "/api/v1/packages/customer-packages": {
      "get": {
        "tags": [
          "Customer"
        ],
        "description": "Packages",
        "requestBody": {
          "content": {
            "application/json": null
          }
        },
        "responses": {
          "200": {
            "message": "Packages"
          }
        }
      }
    },
    "/api/v1/packages/test-packages": {
      "get": {
        "tags": [
          "Customer"
        ],
        "description": "Test Packages",
        "requestBody": {
          "content": {
            "application/json": null
          }
        },
        "responses": {
          "200": {
            "message": "Packages"
          }
        }
      }
    },
...
  },
...
}